backend/requirements.txt
fastapi
supabase
python-dotenv
openai==1.63.0
requests
